+++ Bootstrapper/FodyWeavers.xml
@@ -1,3 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Weavers>
+  <Costura>
+    <ExcludeAssemblies>
+    </ExcludeAssemblies>
+  </Costura>
 </Weavers>
